bench: split into a separate llmman-bench binary
Like ollama-bench, this is now its own binary (cargo build --bin
llmman-bench), not an llmman subcommand, and it doesn't auto-start
llmman serve — start that first, matching ollama-bench's own
assumption. Required splitting the crate into a lib (src/lib.rs) plus
two bins so llmman-bench can reuse daemon/shortnames.

Also addresses review feedback from PR #269:
- Unload a model even if its benchmark run errored out, not just on
  success (wrap the per-model work, unload, then propagate any error).
- Fix a variant-collision bug: epoch 1000 attempt 0 and epoch 0
  attempt 1 both produced variant 1000 under the old epoch +
  attempt*1000 formula, so their prompts could match and defeat the
  KV-cache-busting this was meant to guarantee. Now
  epoch*(retries+1)+attempt, which never repeats.
- Verbose output goes to stderr, not the results writer, so --verbose
  can't corrupt --format csv.

Also shortened comments throughout for signal-to-noise.

All lib/bin unit tests pass (141 + 13); fmt and clippy clean.
