+++ src/cmd/mod.rs
@@ -1,4 +1,3 @@
-pub mod bench;
 pub mod build;
 pub mod inspect;
 pub mod launch;
